==> 2-feature-engineering.ipynb <==
import random, pandas as pd
from datetime import datetime as dt

data_folder = 'G:/Geoff/Data/housing-production'
data_file = 'merged.csv'
save_file = 'cleaned_features.csv'

## Load the data and inspect its shape

# load sales+assessor data
dtypes = {'ucb_geo_id' : str,
          'sr_date_transfer' : str}
df_all = pd.read_csv('{}/{}'.format(data_folder, data_file), low_memory=False, dtype=dtypes)
df_all.shape

df_all.head()

df_all.iloc[0]

# what are the first and last sale dates?
print(df_all['sr_date_transfer'].min())
print(df_all['sr_date_transfer'].min())

## Divide data into before/after

Let dummy variable `after` represent if observation at year1 instead of year0.

# define which years we will look at sales records for, to divide data into before/after
sale_year = 'sr_date_transfer_year'
year0 = 2009
year1 = 2014

# parse date string to datetime object and extract year
# have to do this instead of nice fast pd.to_datetime() style because we 
# have to handle errors from fields with bad date data
def extract_date(date, format='%Y%m%d'):
    try:
        #return pd.to_datetime(date, format=format)
        return dt.strptime(date, format)
    except:
        print('cannot convert "{}"'.format(date))
        
df_all['sr_date_transfer'] = df_all['sr_date_transfer'].map(extract_date)
df_all['sr_date_transfer_year'] = df_all['sr_date_transfer'].map(lambda x: x.year)
df_all.drop(labels='sr_date_transfer', axis='columns', inplace=True)

df = pd.DataFrame(df_all[df_all[sale_year].isin([year0, year1])])
df[sale_year] = df[sale_year].astype(int)
df['after'] = (df[sale_year]==year1).astype(int)

# how many observations do we have at each year in each county?
df.groupby(['MM_MUNI_NAME',sale_year])['ucb_price_sqft_adj'].count().unstack()

# how much did mean sales prices change in each county from year0 to year1?
diff = df.groupby(['MM_MUNI_NAME',sale_year])['ucb_price_sqft_adj'].mean().unstack().round(2)
diff['pct_change'] = (diff[year1] / diff[year0])
diff

# how many rows in data set now?
len(df)

## Divide data into treatment/control

Assign to treatment group if property had at least *n* new housing units built within *m* meters of it between year0 and year1.

# proximity to new building construction


# until we have housing units built var, just assign to groups randomly
df['treat'] = df.index.map(lambda x: random.randint(0, 1))

# or say that 'single family residence' is the treatment group
df['treat'] = (df['USE_CODE_STD']=='RSFR').astype(int)

## Feature engineering

# how old was the building at year1
df['age'] = year1 - df['SA_YR_BLT']

# convert SA_SITE_MAIL_SAME to ternary variable
# to make it a binary dummy, we'll have to handle the unknown "U" cases by dropping them
mapping = {'N' : 0,
           'Y' : 1,
           'U' : 2}
df['SA_SITE_MAIL_SAME'] = df['SA_SITE_MAIL_SAME'].replace(mapping)
mask = ~(df['SA_SITE_MAIL_SAME']==2)
df = df.loc[mask]

# dummy variable for if sale was after construction (1 if yes, 0 if no)
df['sale_after_construction'] = df['sr_date_transfer_year'] > df['SA_YR_BLT']
df['sale_after_construction'] = df['sale_after_construction'].astype(int)

# adjust bathrooms and bedrooms by the number of units?
# Sam created a flag for condo subdivision in the sales dataset that may help with the units issue from the assessor
# but handle zero units to not get inf values
bedrooms_per_unit = df['SA_NBR_BEDRMS'] / df['SA_NBR_UNITS']
bedrooms_per_unit.head()

# retain only certain use codes, and convert to dummies


# identify subsidized units (and drop them?)


# Other independent variables we'll add in later are at the neighborhood level, which we can merge in with the fips numbers.


df.iloc[0]

df.to_csv('{}/{}'.format(data_folder, save_file), index=False, encoding='utf-8')

## Exploring...

mask_0units = df['SA_NBR_UNITS']==0
df[mask_0units]['USE_CODE_STD'].value_counts()

DataQuick Residential Use Codes

```
RAPT	Multi-Family Res (5+ Units)
RCON	Condominium, PUD
RCOO	Cooperative
RDUP	Duplex
RMFD	Multi-Family Dwelling (2-4 Units)
RMOB	Mobile / Manufactured Home
RMSC	Miscellaneous Residential
RQUA	Quadraplex
RSFR	Single Family Residence
RTIM	Timeshare
RTRI	Triplex
```

df[df['USE_CODE_STD']=='RSFR']['SA_NBR_UNITS'].value_counts().sort_index().tail()

counts = df.groupby(['MM_MUNI_NAME', 'USE_CODE_STD'])['sr_property_id'].count()

df_0br = df[df['SA_NBR_BEDRMS'] <= 0]
df_0ba = df[df['SA_NBR_BATH_DQ'] <= 0]
df_0units = df[df['SA_NBR_UNITS'] <= 0]

counts_0br = df_0br.groupby(['MM_MUNI_NAME', 'USE_CODE_STD'])['sr_property_id'].count()
counts_0ba = df_0ba.groupby(['MM_MUNI_NAME', 'USE_CODE_STD'])['sr_property_id'].count()
counts_0units = df_0units.groupby(['MM_MUNI_NAME', 'USE_CODE_STD'])['sr_property_id'].count()

df_counts = pd.DataFrame({'count_observations':counts,
                          'with_0br':counts_0br/counts,
                          'with_0ba':counts_0ba/counts,
                          'with_0units':counts_0units/counts}).fillna(0)

df_counts.to_csv('count-zeroes.csv', index=True, encoding='utf-8')

df[(df['USE_CODE_STD']=='RSFR') & (df['SA_NBR_UNITS']>=5000)].sort_values('SA_NBR_UNITS')

# why are there many RSFRs with 5,000+ units?
# drop all mobile and timeshare?

multi_unit_uses = ['RAPT', 'RCON', 'RCOO', 'RDUP', 'RMFD', 'RMSC', 'RQUA', 'RTRI']
mask_multi_unit = df['USE_CODE_STD'].isin(multi_unit_uses)

df_filtered = df[~(mask_0units & mask_multi_unit)]

print(len(df))
print(len(df_filtered))

df[mask_0units].iloc[:5].T
