==> 3-modeling.ipynb <==
import pandas as pd
import statsmodels.formula.api as smf

data_folder = 'G:/Geoff/Data/housing-production'
data_file = 'cleaned_features.csv'

## Load the data and inspect its shape

# load sales+assessor data
dtypes = {'ucb_geo_id' : str,
          'sr_date_transfer' : str}
df = pd.read_csv('{}/{}'.format(data_folder, data_file), low_memory=False, dtype=dtypes)
df.shape

df.head()

df.iloc[0]

## Inspect the data

df['sr_date_transfer_year'].min()

## Model

formula='ucb_price_sqft_adj ~ treat * after'
simple_model = smf.ols(formula=formula, data=df)
result = simple_model.fit()
print(result.summary())

#### Now estimate a full model by adding covariates

  - building age
  - number of units
  - number of bedrooms
  - number of bathrooms
  - building square feet
  - land use
  - census tract population density
  - census tract percent white
  - census tract median household income

formula = 'ucb_price_sqft_adj ~ treat * after + age'
full_model = smf.ols(formula=formula, data=df)
result = full_model.fit()
print(result.summary())
